+++ docs/source/changelog.rst
@@ -31,9 +31,11 @@ Ksconf v0.11.5 (2023-08-25)
 *  Add new :py:func:`~ksconf.conf.parser.parse_string` function to enable simple parsing from a string.
    (I'm not sure why it took me so long to add this; it's so simple; and I've looked for this function dozens of times over the years, and always came up with a new StringIO workaround.)
 *  App Manifest changes:
+
    *  App manifest file format was updated to v2 which reports file mode as as familiar octal (string) instead of an integer.
       This is easier to read in most cases.
    *  Add API level improvements to support reading/writing manifests when the archive is using a temporary filename.
+
 *  Improved various error messages, minor optimizations, and some minor security improvements.
 
 Ksconf v0.11.4 (2023-06-09)